+++ Riot/Modules/Secrets/Recover/RecoverWithKey/SecretsRecoveryWithKeyViewAction.swift
@@ -18,6 +18,7 @@ import Foundation
 
 /// SecretsRecoveryWithKeyViewController view actions exposed to view model
 enum SecretsRecoveryWithKeyViewAction {
-    case recover    
+    case recover
+    case resetSecrets
     case cancel
 }