+++ Riot/Modules/SetPinCode/EnterPinCode/EnterPinCodeViewModel.swift
@@ -25,6 +25,7 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
     // MARK: Private
 
     private let session: MXSession?
+    private var originalViewMode: SetPinCoordinatorViewMode
     private var viewMode: SetPinCoordinatorViewMode
     
     private var firstPin: String = ""
@@ -40,13 +41,16 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
     weak var viewDelegate: EnterPinCodeViewModelViewDelegate?
     weak var coordinatorDelegate: EnterPinCodeViewModelCoordinatorDelegate?
     private let pinCodePreferences: PinCodePreferences
+    private let localAuthenticationService: LocalAuthenticationService
     
     // MARK: - Setup
     
     init(session: MXSession?, viewMode: SetPinCoordinatorViewMode, pinCodePreferences: PinCodePreferences) {
         self.session = session
+        self.originalViewMode = viewMode
         self.viewMode = viewMode
         self.pinCodePreferences = pinCodePreferences
+        self.localAuthenticationService = LocalAuthenticationService(pinCodePreferences: pinCodePreferences)
     }
     
     // MARK: - Public
@@ -68,7 +72,7 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
             //  go back to first state
             self.update(viewState: .choosePin)
         case .forgotPinAlertResetAction:
-            self.coordinatorDelegate?.enterPinCodeViewModelDidCompleteWithReset(self)
+            self.coordinatorDelegate?.enterPinCodeViewModelDidCompleteWithReset(self, dueToTooManyErrors: false)
         case .forgotPinAlertCancelAction:
             //  no-op
             break
@@ -84,16 +88,41 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
                 return
             } else {
                 currentPin.removeLast()
+                
+                //  switch to setPin if blocked
+                if viewMode == .notAllowedPin {
+                    //  clear error UI
+                    update(viewState: viewState(for: originalViewMode))
+                    //  switch back to original flow
+                    viewMode = originalViewMode
+                }
             }
         } else {
             //  a digit tapped
+            
+            //  switch to setPin if blocked
+            if viewMode == .notAllowedPin {
+                //  clear old pin first
+                currentPin.removeAll()
+                //  clear error UI
+                update(viewState: viewState(for: originalViewMode))
+                //  switch back to original flow
+                viewMode = originalViewMode
+            }
+            //  add new digit
             currentPin += String(tag)
             
             if currentPin.count == pinCodePreferences.numberOfDigits {
                 switch viewMode {
-                case .setPin:
+                case .setPin, .setPinAfterLogin, .setPinAfterRegister:
                     //  choosing pin
                     if firstPin.isEmpty {
+                        //  check if this PIN is allowed
+                        if pinCodePreferences.notAllowedPINs.contains(currentPin) {
+                            viewMode = .notAllowedPin
+                            update(viewState: .notAllowedPin)
+                            return
+                        }
                         //  go to next screen
                         firstPin = currentPin
                         currentPin.removeAll()
@@ -114,6 +143,12 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
                     if currentPin != pinCodePreferences.pin {
                         //  no match
                         numberOfFailuresDuringEnterPIN += 1
+                        pinCodePreferences.numberOfPinFailures += 1
+                        if viewMode == .unlock && localAuthenticationService.shouldLogOutUser() {
+                            //  log out user
+                            self.coordinatorDelegate?.enterPinCodeViewModelDidCompleteWithReset(self, dueToTooManyErrors: true)
+                            return
+                        }
                         if numberOfFailuresDuringEnterPIN < pinCodePreferences.allowedNumberOfTrialsBeforeAlert {
                             DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
                                 self.viewDelegate?.enterPinCodeViewModel(self, didUpdateViewState: .wrongPin)
@@ -126,6 +161,9 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
                         }
                     } else {
                         //  match
+                        //  we can use biometrics anymore, if set
+                        pinCodePreferences.canUseBiometricsToUnlock = nil
+                        pinCodePreferences.resetCounters()
                         //  complete with a little delay
                         DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
                             self.coordinatorDelegate?.enterPinCodeViewModelDidComplete(self)
@@ -139,10 +177,23 @@ final class EnterPinCodeViewModel: EnterPinCodeViewModelType {
         }
     }
     
+    private func viewState(for mode: SetPinCoordinatorViewMode) -> EnterPinCodeViewState {
+        switch mode {
+        case .setPin:
+            return .choosePin
+        case .setPinAfterLogin:
+            return .choosePinAfterLogin
+        case .setPinAfterRegister:
+            return .choosePinAfterRegister
+        default:
+            return .inactive
+        }
+    }
+    
     private func loadData() {
         switch viewMode {
-        case .setPin:
-            update(viewState: .choosePin)
+        case .setPin, .setPinAfterLogin, .setPinAfterRegister:
+            update(viewState: viewState(for: viewMode))
             self.viewDelegate?.enterPinCodeViewModel(self, didUpdateCancelButtonHidden: pinCodePreferences.forcePinProtection)
         case .unlock:
             update(viewState: .unlock)