+++ Riot/Modules/SetPinCode/EnterPinCode/EnterPinCodeViewState.swift
@@ -21,6 +21,9 @@ import Foundation
 /// EnterPinCodeViewController view state
 enum EnterPinCodeViewState {
     case choosePin              //  creating pin for the first time, enter for first
+    case choosePinAfterLogin    //  creating pin for the first time, after login, enter for first
+    case choosePinAfterRegister //  creating pin for the first time, after registration, enter for first
+    case notAllowedPin          //  creating pin for the first time, provided pin is not allowed
     case confirmPin             //  creating pin for the first time, confirm
     case pinsDontMatch          //  pins don't match
     case unlock                 //  after pin has been set, enter pin to unlock